fix(media-encoder): upload retries and worker fixes

+++ workers/media-encoder/src/services/chunk-uploader.ts
@@ -4,7 +4,7 @@ import { Injector } from '@furystack/inject'
 import { ScopedLogger } from '@furystack/logging'
 import { ObservableValue, sleepAsync } from '@furystack/utils'
 import Semaphore from 'semaphore-async-await'
-import got from 'got'
+import got, { Response } from 'got'
 import { media } from '@common/models'
 import FormData from 'form-data'
 
@@ -66,40 +66,10 @@ export class ChunkUploader {
         form.append('mode', this.options.mode)
         form.append('chunk', createReadStream(path))
         form.append('percent', percent)
-        const response = await got(this.options.uploadPath, {
-          method: 'POST',
-          body: form,
-          encoding: 'utf-8',
-          cache: false,
-          agent: false,
-          timeout: {
-            socket: 10000,
-          },
-          retry: { limit: this.options.retries, methods: ['POST'] },
-          hooks: {
-            beforeRetry: [
-              async (_options, error, retrycount) => {
-                await this.logger.information({
-                  message: `Chunk upload has been failed, will retry (${retrycount}/${this.options.retries})...`,
-                  data: { codec: this.options.codec, mode: this.options.mode, fileName, error, retrycount },
-                })
-              },
-            ],
-
-            beforeError: [
-              async (e) => {
-                await this.logger.information({
-                  message: 'Chunk upload has been failed, giving up...',
-                  data: { codec: this.options.codec, mode: this.options.mode, fileName, e },
-                })
-                return e
-              },
-            ],
-          },
-        })
+        const response = await this.uploadWithRetries(form, fileName)
         if (JSON.parse(response.body).success) {
-          await this.logger.verbose({ message: `Chunk '${fileName}' uploaded - ${percent.toFixed(2)}% done.` })
-          if (!fileName.includes('dash.mpd')) {
+          await this.logger.verbose({ message: `Chunk '${fileName}' uploaded - ${(percent || 0).toFixed(2)}% done.` })
+          if (!fileName.includes('dash.mpd') && !fileName.includes('init-stream')) {
             await promises.unlink(path)
           }
         } else {
@@ -121,5 +91,40 @@ export class ChunkUploader {
 
   constructor(private readonly options: ChunkWatcherOptions) {
     this.logger = options.injector.logger.withScope('ChunkUploader')
+  }
+
+  private async uploadWithRetries(form: FormData, fileName: string): Promise<Response<string>> {
+    let retries = 0
+    do {
+      try {
+        return await got(this.options.uploadPath, {
+          method: 'POST',
+          body: form,
+          encoding: 'utf-8',
+          cache: false,
+          agent: false,
+          timeout: {
+            socket: 2000,
+          },
+          retry: {
+            methods: ['POST'],
+          },
+        })
+      } catch (error) {
+        if (retries >= this.options.retries) {
+          await this.logger.error({
+            message: 'Chunk upload has been failed, giving up...',
+            data: { codec: this.options.codec, mode: this.options.mode, fileName, error },
+          })
+          throw error
+        } else {
+          await this.logger.warning({
+            message: `Chunk upload has been failed, will retry (${retries}/${this.options.retries})...`,
+            data: { codec: this.options.codec, mode: this.options.mode, fileName, error, retries },
+          })
+        }
+      }
+    } while (retries++ < this.options.retries)
+    throw Error('No response')
   }
 }
